move server out of generator package
