+++ server/server.go
@@ -1,4 +1,4 @@
-package generator
+package server
 
 import (
 	"github.com/sasswart/gin-in-a-can/model"
@@ -6,13 +6,13 @@ import (
 	"github.com/sasswart/gin-in-a-can/route"
 )
 
-type ServerInterface struct {
+type Interface struct {
 	Routes []route.Route
 	Models []model.Model
 }
 
-func NewServerInterface(openAPIFile string, apiSpec openapi.OpenAPI) ServerInterface {
-	serverInterface := ServerInterface{
+func NewServerInterface(openAPIFile string, apiSpec openapi.OpenAPI) Interface {
+	serverInterface := Interface{
 		Routes: route.NewRoutes(openAPIFile, apiSpec),
 		Models: model.NewModels(openAPIFile, apiSpec),
 	}
